Does not remove board from project when project have only one board

+++ app/services/boards/destroy_service.rb
@@ -1,6 +1,8 @@
 module Boards
   class DestroyService < Boards::BaseService
     def execute(board)
+      return false if project.boards.size == 1
+
       board.destroy
     end
   end
